requirements.txt
requests==2.23.0
discord.py==1.3.3
siaskynet==1.0.2